Task 9 type of variables changed

+++ task-9/task-9.cpp
@@ -4,7 +4,7 @@ using namespace std;
 
 int main()
 {
-    int max, currentValue = 1;
+    double max, currentValue = 1.0;
 
     cout << "Enter a maximum number: ";
     cin >> max;
